+++ R/check_groups.R
@@ -1,0 +1,12 @@
+check_groups <- function(files,
+                         groups){
+    if(is.null(groups)){
+        groups <- rep("all",length(files))
+    } else {
+        if(length(groups)!=length(files)){
+            stopper("groups must be the same length as files or NULL.")
+        }
+    }
+    return(stats::setNames(unique(groups),
+                           unique(groups)))
+}